% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{color_palette}
\alias{color_palette}
\title{Generate a color palette from an image.}
\usage{
color_palette(file, length = 10, exclude = c("#FFFFFF", "#000000"),
  normalize = TRUE)
}
\arguments{
\item{file}{The path or url to an image file (urls should begin with http(s):// or ftp(s)://).}

\item{length}{The number of colors to return.}

\item{exclude}{Colors to exclude from palette. Excludes white and black by default.}

\item{normalize}{Whether or not to apply \code{\link[magick]{image_normalize}}}
}
\value{
A character vector containing a color palette.
}
\description{
Generate a color palette from an image.
}
