% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describeData.R
\name{describe_data}
\alias{describe_data}
\title{Return a sentence describing the given data}
\usage{
describe_data(data, dimension, metric, subject = "users",
  action = "made up", label = "\%", object = "of total traffic")
}
\arguments{
\item{data}{A dataframe to summarize.}

\item{dimension}{The dimension (or grouping variable).}

\item{metric}{The data to summarize.}

\item{subject}{The subject of the sentence (e.g. users, pageviews, etc.)}

\item{action}{The action of the subject (e.g. made up, spent, etc.)}

\item{label}{The label for the metric (e.g. \%, minutes, etc.)}

\item{object}{Noun phrase to include in predicate (e.g. "of total traffic", "of all users", etc.)}
}
\value{
A sentence describing the data provided.
}
\description{
Return a sentence describing the given data
}
