% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_comparison_data.R
\name{ga_comparison_data}
\alias{ga_comparison_data}
\title{Get Google Analytics Comparison Data}
\usage{
ga_comparison_data(viewId, date_range = c(Sys.Date(), Sys.Date() - 1),
  metrics = c("users", "newUsers"), dimensions = c("date",
  "deviceCategory"), anti_sample = TRUE)
}
\arguments{
\item{viewId}{The ID of the Google Analytics View to retrieve data from.}

\item{date_range}{The date range to retrieve data from. Takes a vector of two dates
with the format "YYYY-MM-DD"}

\item{metrics}{Metrics to retrieve from Google Analytics}

\item{dimensions}{Dimensions to retrieve from Google Analytics}

\item{anti_sample}{Set to \code{TRUE} to disallow sampling.}
}
\value{
A list object with data for the current date range,
   the same date range 1 year prior, and the previous period.
}
\description{
A wrapper function around \code{googleAnalyticsR::google_analytics()} that returns
a list with data from three time periods: the time period requested by \code{date_range},
the time period 1 year prior to \code{date_range} and the time period immediately preceding
 \code{date_range}.
}
