% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nps_oppy.R
\name{nps_oppy}
\alias{nps_oppy}
\title{Determine NPS Opportunity for Unique Values of Selected Attribute}
\usage{
nps_oppy(survey_data, group_var, opp_var)
}
\arguments{
\item{survey_data}{Raw survey data. Needs to be one row per survey with the nps question
in a numeric column called nps_question.}

\item{group_var}{Column to group on for baseline NPS calculation.}

\item{opp_var}{Column with attributes that you want to test for opportunity.}
}
\value{
Data frame with baseline NPS and how much NPS would increase by if a given
attribute had perfect NPS scores.
}
\description{
Calculate how much NPS would increase if each distinct value of an attribute had a perfect
NPS score. Optionally choose a grouping column to do it by the values of that column.
}
\details{
The calculation across an entire sample of surveys is simple. A customer experience
manager may want to calculate CXi across many different dimensions and filtering
in different ways; the functions in this package utilize the tidy framework to streamline
calculating CXi along as many dimensions as desired.
}
\examples{
nps_question <- sample(10, 100, replace = TRUE)
nps_date <- rep(seq.Date(from = as.Date("2019-01-01"), to = as.Date("2019-01-10"), by = "day"), 10)
nps_group <- rep(c("a", "b", "c", "d"), 25)
nps_attr <- rep(c("alpha", "beta", "chi", "delta"), 25)
df <- data.frame(nps_question, nps_date, nps_group, nps_attr)
# see improvements to overall NPS if each attribute had a perfect score
nps_oppy(df, group_var = NULL, opp_var = nps_attr)
# see improvements to group-level NPS if each attribute had a perfect score
nps_oppy(df, group_var = nps_group, opp_var = nps_attr)

}
