% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyDates.R
\name{pretty_date}
\alias{pretty_date}
\title{Format a Date for Display.}
\usage{
pretty_date(date, show_year = TRUE, show_month = TRUE,
  show_wday = FALSE)
}
\arguments{
\item{date}{The date to be formatted (either a date object
or a string with the format "YYYY-MM-DD").}

\item{show_year}{Whether or not to display year in output. Default: \code{TRUE}.}

\item{show_month}{Whether or not to display month in output. Default: \code{TRUE}.}

\item{show_wday}{Whether or not to display the day of the week
in output. Default: \code{FALSE}.}
}
\value{
A string that represents the date supplied in \code{date}.
}
\description{
Format a Date for Display.
}
\examples{
prettyDate("9-20-2019")
# September 20th, 2019

prettyDate("9-20-2019", show_wday=T)
# Friday, September 20th, 2019

}
