% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyDates.R
\name{pretty_date_range}
\alias{pretty_date_range}
\title{Format the display of a date period.}
\usage{
pretty_date_range(date1, date2, separator = " - ", show_wday = FALSE)
}
\arguments{
\item{date1, date2}{The dates to be formatted (either date objects
or strings with the format "YYYY-MM-DD").}

\item{show_wday}{Whether or not to display the
day of the week. Default: \code{FALSE}
dates. DEFAULT: \code{" - "}}
}
\value{
A string with both dates provided formatted.
}
\description{
Format the display of a date period.
}
\examples{
prettyDateRange("9-20-2019", "09-21-2019", show_wday=FALSE)
# outputs "September 20th - 21st, 2019"

prettyDateRange("9-20-2019", "10-20-2019", show_wday=FALSE)
# outputs "September 20th - October 20th, 2019"

prettyDateRange("9-20-2019", "10-20-2020", show_wday=FALSE)
# outputs "September 20th, 2019 - October 20th, 2020"

}
